logic/ is favorite

+++ app/src/main/java/ru/spartak/surfandroidschool/data/database/picture_database/entity/PictureEntity.kt
@@ -10,6 +10,6 @@ data class PictureEntity(
     val title: String,
     val content: String,
     val photoUrl: String,
-    val publicationDate: Int,
+    val publicationDate: Long,
     val isFavorite:Boolean,
 )
